requirements.txt
absl-py==2.0.0
accelerate==0.29.2
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.27
botocore==1.34.27
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.28.1
comm==0.2.1
confection==0.1.4
contexttimer==0.3.3
contourpy==1.2.0
ctransformers==0.2.27
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
decord==0.6.0
distlib==0.3.8
docker-pycreds==0.4.0
docstring-parser==0.14.1
easydict==1.11
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.4
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.47.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gprof2dot==2022.7.29
GPUtil==1.4.0
grpcio==1.60.0
h5py==3.10.0
hydra-core==1.3.2
identify==2.5.32
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.6.1
kaggle==1.6.3
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lightning-bolts==0.6.0.post1
lightning-utilities==0.10.0
lit==17.0.6
loguru==0.7.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objaverse==0.1.7
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
pluggy==1.3.0
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
pytest==7.4.4
pytest-profiling==1.7.0
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.2
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
shtab==1.6.5
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
snakeviz==2.2.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2024.1.30
timm
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch @ https://download.pytorch.org/whl/cu117/torch-2.0.1%2Bcu117-cp310-cp310-linux_x86_64.whl#sha256=bb54b705185bea820e6ec6485a25761bc03f689e1a09a37d814d6ea8e276b5bd
torchmetrics==1.2.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
trimesh==4.0.8
triton==2.0.0
typer==0.9.0
typing_extensions==4.8.0
tyro==0.3.38
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
virtualenv==20.25.0
wandb==0.16.2
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.12
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
Werkzeug==3.0.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
